Add a schematic (in Eagle CAD format) file for my custom hardware. I believe this is correct, although I'm rather new to Eagle, so the layout could be better.
There is no board file because I used the "full" Raspberry Pi instead of the one with "just the pins", and you probably don't want to put the Pi *onto* your circuit board. It's usually the other way-round. :-)

Does anyone know how to replace this RPi with the one with just the headers? If so, I'd welcome the contribution! Thanks!